Add Monte Carlo integration example
This code demonstrates two things:
1) That the function triangle-explicit does not yield a uniform random sample,
   so we need to be careful *not* to use a DFS scheduler for amb or we will run
   into the sample problem.
2) That our amb system uses up way too much memory. I tried to sample 500
   points, which means generating an expected 1000 points (because only half of
   all points are in our domain). Each of these 1000 points needs a new env
   with two variables, x and y. So we're storing 2000 floats, and somehow we
   run out of memory? How does that even happen?

I was really hoping I could use this empirical test to figure out if the
amb-amb-require model yields a uniform random sample, but because this runs out
of memory, I can only try it on very small sample sizes -- which are so small
that I can't actually tell if the sample is uniform.

I think we should run a chi-squared test to determine if amb-amb-require is
uniform, except I think there's a big difference between calling a function 100
times and calling try-again 100 times, and I don't know how to do the latter in
an automated fashion.